+++ FloconDesktop/data/remote/src/commonMain/kotlin/com/flocon/data/remote/network/models/FloconNetworkRequestDataModel.kt
@@ -30,7 +30,7 @@ data class FloconNetworkResponseDataModel(
     val floconCallId: String? = null,
     val floconNetworkType: String? = null,
 
-    val durationMs: Long? = null,
+    val durationMs: Double? = null,
     val responseHttpCode: Int? = null, // ex: 200
     val responseContentType: String? = null,
     val responseBody: String? = null,
@@ -43,7 +43,7 @@ data class FloconNetworkResponseDataModel(
 internal fun FloconNetworkResponseDataModel.toDomain(): FloconNetworkResponseOnlyDomainModel? {
     return try {
         val callId = floconCallId!!
-        val durationMs = durationMs ?: 0
+        val durationMs = durationMs ?: 0.0
 
         val response = if (responseError != null) {
             FloconNetworkCallDomainModel.Response.Failure(
